backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pytesseract==0.3.13
pdf2image==1.17.0
pymupdf==1.24.10
faiss-cpu==1.8.0
sentence-transformers==3.1.1
whisper-cpp-python==0.0.15
TTS==0.23.0
python-dotenv==1.0.1
cryptography==43.0.1